Initial leisa2isis app test conversion and gtest
