+++ isis/src/newhorizons/apps/leisa2isis/leisa2isis.h
@@ -1,0 +1,11 @@
+#ifndef leisa2isis_h
+#define leisa2isis_h
+
+#include "Pvl.h"
+#include "UserInterface.h"
+
+namespace Isis{
+  extern void leisa2isis(UserInterface &ui, Pvl *log);
+}
+
+#endif
